Feat: 스크랩 생성, 삭제시 각각 validation 추가
- 생성시 post 가 이미 삭제된 상태라면 PostAlreadyDeletedException 호출
- 삭제시 해당 스크랩이 유저의 것이 아니라면 ScrapNotAuthorizedException 호출

+++ src/main/java/com/justcodeit/moyeo/study/application/scrap/ScrapService.java
@@ -3,7 +3,8 @@ package com.justcodeit.moyeo.study.application.scrap;
 import com.justcodeit.moyeo.study.application.post.exception.PostCannotFoundException;
 import com.justcodeit.moyeo.study.application.scrap.exception.PostAlreadyDeletedException;
 import com.justcodeit.moyeo.study.application.scrap.exception.ScrapCannotFoundException;
-import com.justcodeit.moyeo.study.model.scrap.ScrapQueryDto;
+import com.justcodeit.moyeo.study.application.scrap.exception.ScrapNotAuthorizedException;
+import com.justcodeit.moyeo.study.model.inquiry.ScrapQueryDto;
 import com.justcodeit.moyeo.study.model.post.PostStatus;
 import com.justcodeit.moyeo.study.persistence.Post;
 import com.justcodeit.moyeo.study.persistence.Scrap;
@@ -27,19 +28,22 @@ public class ScrapService {
     Post post = postRepository.findById(postId)
             .orElseThrow(PostCannotFoundException::new);
 
-    isPostDeleted(post);
+    // TODO : http 상태코드? postStatus abnormal?
+    if (post.getPostStatus() == PostStatus.DELETE) {
+      throw new PostAlreadyDeletedException("이미 삭제된 모집글입니다");
+    }
     Scrap scrap = new Scrap(userId, post.getId());
 
     return scrapRepository.save(scrap).getId();
   }
 
   @Transactional
-  public void deleteScrap(Long scrapId) {
+  public void deleteScrap(String userId, Long scrapId) {
     Scrap scrap = findScrap(scrapId);
-    Post post = postRepository.findById(scrap.getPostId())
-            .orElseThrow(PostCannotFoundException::new);
 
-    isPostDeleted(post);
+    if (!scrap.getUserId().equals(userId)) {
+      throw new ScrapNotAuthorizedException();
+    }
     scrapRepository.delete(scrap);
   }
 
@@ -48,15 +52,8 @@ public class ScrapService {
     return scrapRepository.findScrapListByUserId(userId);
   }
 
-  @Transactional(readOnly = true)
   private Scrap findScrap(Long scrapId) {
     return scrapRepository.findById(scrapId)
             .orElseThrow(() -> new ScrapCannotFoundException(String.format("해당 스크랩을 찾을 수 없습니다 : %s", scrapId)));
-  }
-
-  private void isPostDeleted(Post post) {
-    if (post.getPostStatus() == PostStatus.DELETE) {
-      throw new PostAlreadyDeletedException();
-    }
   }
 }
